arm: u8500_hotplug: new algorithm
instead of traditionally hotplugging based on load, this new algorithm should allow better performance for our poor dual-core phone by not offlining cpu1 that much.
if the current load on cpu0 is high(default 50), cpu1 is immediately onlined, no questions asked.
however, if it isn't, a counter is added at every poll when the current load is low.
once the counter reaches 10, it is reset and cpu1 is offlined until the next poll.
this should allow better performance by offlining cpu1 only when the user is really not doing anything much.

also, now the high load threshold and the counter threshold can be tuned.
Change-Id: Ief11ac1d4f9fc52ed818a37e0aeccf2a9f4b99ce